refactor: consolidate types - merge types.ts into types/index.ts, standardize Album interface, clean imports and unused Script
